feat(reservas): implement ReservaEditPage and ReservaHistorialPage
- Added ReservaEditPage.vue for editing reservations with a tabbed interface for main data, services, additional items, and summary.
- Integrated a floating save button and a dialog for creating new clients.
- Updated ReservaFormPage.vue to support both creation and editing modes with a stepper and tab navigation.
- Introduced ReservaHistorialPage.vue to display the change history of reservations with detailed views for each change.
- Enhanced ReservaListPage.vue to include a button for viewing the change history of reservations.
- Updated routing to accommodate new pages and ensure proper navigation.

+++ backend/base/models.py
@@ -33,7 +33,8 @@ class Auditoria(models.Model):
     accion = models.CharField(max_length=1, choices=ACCIONES)
     fecha = models.DateTimeField(auto_now_add=True)
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    # Cambiado a BigInt para soportar IDs temporales grandes
+    object_id = models.PositiveBigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
     datos_anteriores = models.JSONField(null=True, blank=True)
     datos_nuevos = models.JSONField(null=True, blank=True)
